feat: create dir middleware
